Start putting in marker structs
