+++ lib/rs/src/lib.rs
@@ -58,16 +58,10 @@ pub mod transport;
 mod errors;
 pub use errors::*;
 
-use self::protocol::{TInputProtocol, TOutputProtocol};
+mod autogen;
+pub use autogen::*;
 
 /// Result type returned by all rift functions.
 /// As is convention, this is a typedef of `std::result::Result`
 /// with `E` defined as the rift `Error` type.
 pub type Result<T> = std::result::Result<T, self::Error>;
-
-pub trait TThriftClient {
-    fn i_prot(&mut self) -> &mut TInputProtocol;
-    fn o_prot(&mut self) -> &mut TOutputProtocol;
-    fn sequence_number(&self) -> i32;
-    fn increment_sequence_number(&mut self);
-}
